+++ appcontainer/AppStateModel.ts
@@ -24,8 +24,13 @@ export class AppStateModel extends HoistModel {
     suspendData: AppSuspendData;
     accessDeniedMessage: string = 'Access Denied';
 
+    /**
+     * Timestamp when the app first started loading, prior to even JS download/eval.
+     * Read from timestamp set on window within index.html.
+     */
+    readonly loadStarted: number = window['_xhLoadTimestamp'];
+
     private timings: Record<AppState, number> = {} as Record<AppState, number>;
-    private loadStarted: number = window['_xhLoadTimestamp']; // set in index.html
     private lastStateChangeTime: number = this.loadStarted;
 
     constructor() {